Simplify: deduplicate void helpers, clean imports, remove JIRA refs
- Void helpers now delegate to value-returning variants (-20 lines)
- Remove duplicate javadoc on withSynchronizedDatabaseVoid
- Update withSynchronizedDatabase javadoc: precondition not assertion
- Remove JIRA ticket references from comments (stale post-merge)
- Remove unnecessary (long) cast in getRmdByteUsage()
- Use imports instead of FQN in test (Assert, Collections,
BytesStreamingCallback)
- Consolidate endBatchWrite block comments into one-liner

+++ clients/da-vinci-client/src/main/java/com/linkedin/davinci/store/rocksdb/ReplicationMetadataRocksDBStoragePartition.java
@@ -113,8 +113,7 @@ public class ReplicationMetadataRocksDBStoragePartition extends RocksDBStoragePa
 
   public long getRmdByteUsage() {
     return withOpenDatabase(
-        db -> (long) db.getColumnFamilyMetaData(columnFamilyHandleList.get(REPLICATION_METADATA_COLUMN_FAMILY_INDEX))
-            .size());
+        db -> db.getColumnFamilyMetaData(columnFamilyHandleList.get(REPLICATION_METADATA_COLUMN_FAMILY_INDEX)).size());
   }
 
   /**
